setInterval now takes a "long". getInterval returns a long, and the attribute type is a long.
git-svn-id: https://cruisecontrol.svn.sourceforge.net/svnroot/cruisecontrol/trunk/cruisecontrol@1058 878de816-9c12-0410-b5fe-b34e1928e20d

+++ main/src/net/sourceforge/cruisecontrol/Schedule.java
@@ -259,7 +259,7 @@ public class Schedule {
         return timeToEndOfPause + ONE_MINUTE;
     }
 
-    public void setInterval(int intervalBetweenModificationChecks) {
+    public void setInterval(long intervalBetweenModificationChecks) {
         interval = intervalBetweenModificationChecks * ONE_SECOND;
     }
